Issue #1944: Move DelegatingStream to internal namespace and rename to NonDisposableStream.
