+++ src/Microsoft.AspNet.Mvc.Xml/XmlSerializerInputFormatter.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Serialization;
+using Microsoft.AspNet.Mvc.Internal;
 using Microsoft.Net.Http.Headers;
 
 namespace Microsoft.AspNet.Mvc.Xml
@@ -146,7 +147,7 @@ namespace Microsoft.AspNet.Mvc.Xml
         {
             var request = context.ActionContext.HttpContext.Request;
 
-            using (var xmlReader = CreateXmlReader(new DelegatingStream(request.Body)))
+            using (var xmlReader = CreateXmlReader(new NonDisposableStream(request.Body)))
             {
                 var type = GetSerializableType(context.ModelType);
